+++ g2g-plugins/src/webrtcsession.rs
@@ -49,7 +49,7 @@ use g2g_core::{
 };
 
 use crate::filesink::io_err;
-use crate::turn::{self, TurnClient};
+use crate::turn::{self, TurnSet};
 use crate::webrtc_util::{
     add_host_candidate, delete_resource, feed_datagram, ice_restart, post_sdp, select_host_ip,
     send_transmit, trickle_candidates, TricklePatch, TurnConfig, ICE_RESTART_TIMEOUT,
@@ -221,7 +221,7 @@ impl WebRtcSessionSink {
             .map_err(|_| hw())?;
 
         // Gather reflexive / relay candidates and trickle them to the resource.
-        let turn = trickle_candidates(
+        let turn: TurnSet = trickle_candidates(
             &mut rtc,
             &socket,
             &offer_sdp,
@@ -502,7 +502,7 @@ async fn run_session(
     audio_mid: Option<Mid>,
     video_reverse: Option<ReverseChannel>,
     audio_reverse: Option<ReverseChannel>,
-    mut turn: Option<TurnClient>,
+    mut turn: TurnSet,
     resource: Option<String>,
     etag: Option<String>,
     bearer: Option<String>,
@@ -619,10 +619,8 @@ async fn run_session(
                     }
                 }
             }
-            _ = tokio::time::sleep_until(tokio::time::Instant::from_std(refresh_at)), if turn.is_some() => {
-                if let Some(tc) = turn.as_mut() {
-                    let _ = tc.refresh(&socket).await;
-                }
+            _ = tokio::time::sleep_until(tokio::time::Instant::from_std(refresh_at)), if !turn.is_empty() => {
+                turn.refresh_all(&socket).await;
                 refresh_at = Instant::now() + turn::REFRESH_INTERVAL;
             }
             _ = tokio::time::sleep(timeout) => {